requirements.txt
python>=3.11
youtube-search-python==1.6.6
youtube-transcript-api==0.6.2
pytube==15.0.0
langchain-community==0.0.34
langchain-core==0.1.99 
faiss-cpu==1.7.4
sentence-transformers==2.2.2
faster-whisper==0.10.0
ollama==0.1.6
requests==2.32.3
tqdm==4.66.2
python-dotenv==1.0.1
typing_extensions==4.11.0
